chore: rename feature
