--- packages/transport/src/modules/notifications/backbone/BackbonePostNotifications.ts
@@ -1,4 +1,0 @@
-export interface BackbonePostNotificationsRequest {
-    recipients: string[];
-    code: string;
-}
